test(zenanalyze): fix 32-bit usize overflow in spectral_ln_binning_perf
The synthetic-block LCG (b*64+i)*2_654_435 overflows 32-bit usize on i686 (debug
overflow panic in the cross job). Compute in u64. The perf test's correctness
assert (sink.is_finite) is what runs on i686; the perf numbers there are
qemu-meaningless but harmless.

+++ src/simd_math.rs
@@ -321,7 +321,8 @@ mod tests {
             .map(|b| {
                 let mut blk = [0.0f32; 64];
                 for (i, v) in blk.iter_mut().enumerate() {
-                    *v = (((b * 64 + i) * 2_654_435 + 1) % 401) as f32 - 200.0;
+                    // u64 math so the LCG doesn't overflow 32-bit usize on i686.
+                    *v = (((b * 64 + i) as u64 * 2_654_435 + 1) % 401) as f32 - 200.0;
                 }
                 blk
             })
